feat(ui): refact sidebar ui v2
